+++ experiments/dqn_experiment.py
@@ -13,7 +13,7 @@ class DQNExperiment(BaseExperiment):
     def __init__(self):
         super().__init__(name="DqnExperiment", config=get_experiment_config())
         self._config: DQNExperimentConfig = get_experiment_config()
-        self._env = self.build_env_v2(self._config.env_config)
+        self._env = self.build_env(self._config.env_config)
 
     def get_rl_model(self) -> RLAgent:
         rl_model = DQN(env=self._env, config=self._config.dqn_config)
