+++ experiments/experiment.py
@@ -14,9 +14,8 @@ from datetime import datetime
 import wandb
 from wandb.wandb_run import Run
 
-from data_generators.data_generators import DataGenerator
 from env.gelateria import GelateriaState, Gelato
-from env.gelateria_env_v2 import GelateriaEnv_v2 as GelateriaEnv
+from env.gelateria_env import GelateriaEnv
 from env.mask.action_mask import ActionMask
 from env.reward.base_reward import BaseReward
 from models.base_rl_agent import RLAgent
@@ -26,10 +25,6 @@ from utils.enums import Flavour
 
 class BaseExperiment:
     """Base class for experiments."""
-
-    # def get_experiment_config():
-    #     config = ExperimentConfig()
-    #     return config
 
     def __init__(self, name: str, config: ExperimentConfig):
         self._name = name
@@ -82,47 +77,13 @@ class BaseExperiment:
         return wandb.init(project=wandb_config.project, entity=wandb_config.entity, config=run_config,
                           mode=wandb_config.mode, name=run_name)
 
-    def _get_dataset_generator(self):
-        return DataGenerator(
-            config=self._config.data_generation_config, dataloader_config=self._config.dataloader_config
-        )
-
-    def get_supervised_model(self, supervised_model: nn.Module) -> nn.Module:
-        dummy_dataloader, _ = self._get_dataset_generator().get_train_val_dataloaders()
-        input_dim = next(iter(dummy_dataloader))["public_obs"].shape[-1]
-        model = supervised_model(input_dim=input_dim, name="mlp_sales", config=self._config.net_config)
-        model.load()
-        return model
-
-    def build_env(self, init_state: GelateriaState, supervised_model: nn.Module, reward: BaseReward,
-                  action_mask_fn: Optional[ActionMask] = None, restock_fct: Optional[Callable] = None):
-        if not isinstance(init_state, GelateriaState):
-            raise ValueError("init_state must be of type GelateriaState")
-        if not callable(supervised_model):
-            raise ValueError("supervised_model must be callable")
-        if not isinstance(reward, BaseReward):
-            raise ValueError("reward must be a subclass of BaseReward")
-
-        self._init_state = init_state
-        self._supervised_model = self.get_supervised_model(supervised_model)
-        self._reward = reward
-        self._action_mask_fn = action_mask_fn
-
-        env = GelateriaEnv(init_state=init_state,
-                           sales_model=self._supervised_model,
-                           reward=reward,
-                           mask_fn=action_mask_fn,
-                           restock_fct=restock_fct)
-
-        return env
-
     @staticmethod
-    def build_env_v2(config: EnvConfig):
+    def build_env(config: EnvConfig):
         from models.sales.sales_prediction_models import GenericSalesPredictionModel
         from models.sales.base_sales_models import CombinedMLPLogBaseSalesModel
         from models.sales.sales_uplift_models import GammaSalesUpliftModel
         from models.sales.dataset import transform_gym_inputs
-        from env.gelateria_env_v2 import GelateriaEnv_v2
+        from env.gelateria_env import GelateriaEnv
         from env_wrapper.wrapper import DefaultGelatoEnvWrapper
 
         sales_model = GenericSalesPredictionModel(
@@ -137,14 +98,14 @@ class BaseExperiment:
         else:
             init_state = BaseExperiment._get_experiment_init_state()
 
-        env = GelateriaEnv_v2(init_state=init_state,
-                              sales_model=sales_model,
-                              reward=config.reward_fn,
-                              mask_fn=config.action_mask_fn,
-                              days_per_step=config.days_per_step,
-                              end_date=config.end_date,
-                              max_steps=config.max_steps,
-                              restock_fct=config.restock_fn)
+        env = GelateriaEnv(init_state=init_state,
+                           sales_model=sales_model,
+                           reward=config.reward_fn,
+                           mask_fn=config.action_mask_fn,
+                           days_per_step=config.days_per_step,
+                           end_date=config.end_date,
+                           max_steps=config.max_steps,
+                           restock_fct=config.restock_fn)
 
         env_wrap = DefaultGelatoEnvWrapper(env)
         return env_wrap
