+++ experiments/mbpo_sac_discrete_experiment.py
@@ -14,7 +14,7 @@ class MbpoSacDiscreteExperiment(BaseExperiment):
     def __init__(self):
         super().__init__(name="MbpoSacDiscreteExperiment", config=get_experiment_config())
         self._config: MBPOExperimentConfig = get_experiment_config()
-        self._env = self.build_env_v2(self._config.env_config)
+        self._env = self.build_env(self._config.env_config)
 
     def get_agent(self) -> RLAgent:
         agent = SACDiscrete(env=self._env, config=self._config.sac_config)
