+++ experiments/rule_based_agent_experiment.py
@@ -13,7 +13,7 @@ class RuleBasedAgentExperiment(BaseExperiment):
     def __init__(self):
         super().__init__(name="RuleBasedAgentExperiment", config=get_experiment_config())
         self._config: RuleBasedAgentExperimentConfig = get_experiment_config()
-        self._env = self.build_env_v2(self._config.env_config)
+        self._env = self.build_env(self._config.env_config)
 
     def get_rl_model(self) -> RLAgent:
         rl_model = RuleBasedAgent(env=self._env, config=self._config.rule_based_agent_config)
